Keeper: address review on the SessionID routing tests
Four review points from antonio2368 on PR #114401.

The SessionID correlation check in KeeperRequestDispatcher::onCommit now uses
assert_cast instead of dynamic_cast plus two nullptr branches. The preceding
opnum check makes ZooKeeperSessionIDRequest the only possible type of the
committed request, and the matching head is one too because both carry the
reserved session id -1, which only KeeperDispatcher::getSessionID pushes.
assert_cast is a static_cast in release builds and typeid-checked under
DEBUG_OR_SANITIZER_BUILD, so the impossible-type branches (which would have
silently stalled the batch) are gone.

The seven per-test friend declarations across KeeperRequestDispatcher.h,
KeeperRequestDispatcherOld.h and KeeperDispatcher.h are replaced by one test
accessor per class, following TemporaryPartsTestAccessor. That also removes the
seven forward declarations the gtest-generated names needed, and adding a test
no longer touches a production header.

The tests move into gtest_coordination.cpp and the standalone
gtest_keeper_session_id_routing.cpp is deleted.

SessionIDRouterIsRequired is removed as requested. The production guards that
reject an empty router stay in both dispatcher constructors; only the assertion
on them is gone, along with the DEBUG_OR_SANITIZER_BUILD conditional it needed.
