+++ src/Coordination/KeeperRequestDispatcher.h
@@ -12,11 +12,6 @@
 extern template class NonblockingBoundedQueue<DB::KeeperRequestForSession>;
 extern template class NonblockingBoundedQueue<DB::KeeperResponseForSession>;
 
-/// White-box unit tests that seed in_flight_batches (befriended inside KeeperRequestDispatcher).
-class KeeperDispatcher_SessionIDCommitCorrelation_Test;
-class KeeperDispatcher_SessionIDErrorReachesWaiter_Test;
-class KeeperDispatcher_SessionIDErrorReachesRealWaiter_Test;
-
 namespace DB
 {
 
@@ -129,9 +124,7 @@ public:
     void onResponseDeallocated(const Coordination::ZooKeeperResponse & response);
 
 private:
-    friend class ::KeeperDispatcher_SessionIDCommitCorrelation_Test;
-    friend class ::KeeperDispatcher_SessionIDErrorReachesWaiter_Test;
-    friend class ::KeeperDispatcher_SessionIDErrorReachesRealWaiter_Test;
+    friend class KeeperRequestDispatcherTestAccessor;
 
     /// Suppose we get a write request from some session and put it in batch B and send that
     /// batch to leader. While B is still in flight, we get a read request from the same session.
