+++ src/Coordination/tests/gtest_coordination.cpp
@@ -8,6 +8,9 @@
 #include <Coordination/SummingStateMachine.h>
 #include <Coordination/KeeperContext.h>
 #include <Coordination/KeeperCommon.h>
+#include <Coordination/KeeperDispatcher.h>
+#include <Coordination/KeeperRequestDispatcher.h>
+#include <Coordination/KeeperRequestDispatcherOld.h>
 #include <Coordination/KeeperServer.h>
 #include <Coordination/KeeperConstants.h>
 #include <Coordination/KeeperStorage.h>
@@ -23,13 +26,17 @@
 
 #include <Coordination/LoggerWrapper.h>
 
+#include <Common/ZooKeeper/KeeperException.h>
 #include <Common/ZooKeeper/ZooKeeperCommon.h>
 #include <Common/ZooKeeper/ZooKeeperIO.h>
 #include <Common/logger_useful.h>
 #include <Common/Exception.h>
+#include <Common/FailPoint.h>
+#include <Common/scope_guard_safe.h>
 
 #include <Poco/Util/XMLConfiguration.h>
 
+#include <future>
 #include <limits>
 #include <sstream>
 
@@ -1255,6 +1262,367 @@ TEST(CoordinationRequestSize, WriteRejectsRequestOverInt32)
     HugeRequest request;
     DB::WriteBufferFromNuraftBuffer wbuf;
     EXPECT_THROW(request.write(wbuf, false, false), Coordination::Exception);
+}
+
+namespace DB
+{
+
+/// The in-flight batch queue and the SessionID error path are private, so the tests below reach them
+/// through these friend accessors.
+class KeeperRequestDispatcherTestAccessor
+{
+public:
+    /// Puts one request in a fresh in-flight batch, the way dispatchThread would, and returns its index.
+    static size_t seedInFlightBatch(KeeperRequestDispatcher & dispatcher, const KeeperRequestForSession & request)
+    {
+        size_t batch_idx = dispatcher.tail_idx.load();
+        auto & batch = dispatcher.in_flight_batches[batch_idx % dispatcher.in_flight_batches.size()];
+        batch.requests = {request};
+        batch.activate({});
+        dispatcher.tail_idx.store(batch_idx + 1);
+        return batch_idx;
+    }
+
+    static size_t committedRequests(KeeperRequestDispatcher & dispatcher, size_t batch_idx)
+    {
+        return dispatcher.in_flight_batches[batch_idx % dispatcher.in_flight_batches.size()].committed_requests;
+    }
+
+    static size_t headIdx(const KeeperRequestDispatcher & dispatcher) { return dispatcher.head_idx.load(); }
+
+    static void dropInFlightRequests(KeeperRequestDispatcher & dispatcher) { dispatcher.dropInFlightRequests(); }
+};
+
+class KeeperRequestDispatcherOldTestAccessor
+{
+public:
+    static void addErrorResponses(
+        KeeperRequestDispatcherOld & dispatcher,
+        const KeeperRequestsForSessions & requests_for_sessions,
+        Coordination::Error error,
+        bool may_have_dependent_reads)
+    {
+        dispatcher.addErrorResponses(requests_for_sessions, error, may_have_dependent_reads);
+    }
+};
+
+class KeeperDispatcherTestAccessor
+{
+public:
+    static void setKeeperContext(KeeperDispatcher & dispatcher, KeeperContextPtr keeper_context)
+    {
+        dispatcher.keeper_context = std::move(keeper_context);
+    }
+
+    static void setServer(KeeperDispatcher & dispatcher, std::unique_ptr<KeeperServer> server)
+    {
+        dispatcher.server = std::move(server);
+    }
+
+    static KeeperServer * server(KeeperDispatcher & dispatcher) { return dispatcher.server.get(); }
+
+    static KeeperSpecialResponseRouter router(KeeperDispatcher & dispatcher)
+    {
+        return [&dispatcher](const KeeperResponseForSession & response)
+        { return dispatcher.tryRouteSpecialResponse(response); };
+    }
+
+    /// Registers a waiter the way getSessionID does, and returns the future a client blocks on.
+    /// Empty if the internal id was already registered.
+    static std::optional<std::future<int64_t>> registerSessionIDWaiter(KeeperDispatcher & dispatcher, int64_t internal_id)
+    {
+        std::lock_guard lock(dispatcher.new_session_id_mutex);
+        auto [it, inserted] = dispatcher.new_session_id_requests.try_emplace(internal_id);
+        if (!inserted)
+            return {};
+        return it->second.get_future();
+    }
+
+    static size_t sessionIDWaiterCount(KeeperDispatcher & dispatcher, int64_t internal_id)
+    {
+        std::lock_guard lock(dispatcher.new_session_id_mutex);
+        return dispatcher.new_session_id_requests.count(internal_id);
+    }
+};
+
+}
+
+namespace
+{
+
+/// A server without a started Raft instance. Enough for onCommit and the error paths, which only
+/// touch in_flight_batches and the response routing.
+struct DispatcherFixture
+{
+    ChangelogDirTest dir{"./session_id_routing_logs"};
+    DB::KeeperContextPtr keeper_context;
+    Poco::AutoPtr<Poco::Util::XMLConfiguration> config;
+    DB::SnapshotsQueue snapshots_queue{1};
+    DB::KeeperSnapshotManagerS3 snapshot_s3;
+    std::unique_ptr<DB::KeeperServer> server;
+    std::unique_ptr<DB::KeeperRequestDispatcher> dispatcher;
+
+    /// Responses the router took, i.e. that did not go to the per-session response queue.
+    std::vector<DB::KeeperResponseForSession> routed;
+
+    DB::KeeperSpecialResponseRouter router()
+    {
+        return [this](const DB::KeeperResponseForSession & response)
+        {
+            if (response.response->getOpNum() != Coordination::OpNum::SessionID)
+                return false;
+            routed.push_back(response);
+            return true;
+        };
+    }
+
+    DispatcherFixture()
+    {
+        std::string xml = R"(<clickhouse><keeper_server>
+            <server_id>1</server_id>
+            <tcp_port>0</tcp_port>
+            <raft_configuration><server>
+                <id>1</id><hostname>localhost</hostname><port>44444</port>
+            </server></raft_configuration>
+        </keeper_server></clickhouse>)";
+        std::stringstream stream(xml); // NOLINT(readability-isolate-declaration)
+        config = new Poco::Util::XMLConfiguration(stream);
+
+        keeper_context = ::makeKeeperContext(false, nullptr);
+        keeper_context->setLogDisk(std::make_shared<DB::DiskLocal>("LogDisk", dir.path));
+        keeper_context->setSnapshotDisk(std::make_shared<DB::DiskLocal>("SnapshotDisk", dir.path));
+        keeper_context->setStateFileDisk(std::make_shared<DB::DiskLocal>("StateFile", dir.path));
+        keeper_context->setLocalLogsPreprocessed();
+
+        server = std::make_unique<DB::KeeperServer>(
+            DB::KeeperConfiguration::loadFromConfig(*config, true),
+            *config,
+            [](DB::KeeperResponseForSession) {},
+            snapshots_queue,
+            keeper_context,
+            snapshot_s3,
+            [](uint64_t, const DB::KeeperRequestForSession &) {});
+
+        dispatcher = std::make_unique<DB::KeeperRequestDispatcher>(server.get(), router());
+    }
+};
+
+DB::KeeperRequestForSession makeSessionIDRequest(int32_t server_id, int64_t internal_id)
+{
+    auto request = std::make_shared<Coordination::ZooKeeperSessionIDRequest>();
+    request->server_id = server_id;
+    request->internal_id = internal_id;
+    request->session_timeout_ms = 10000;
+    /// KeeperDispatcher::getSessionID leaves xid at its default and uses session id -1, so every
+    /// SessionID request in the cluster carries the same (session_id, xid).
+    DB::KeeperRequestForSession request_for_session;
+    request_for_session.request = request;
+    request_for_session.session_id = DB::keeper_internal_get_session_id;
+    return request_for_session;
+}
+
+using RequestDispatcherAccessor = DB::KeeperRequestDispatcherTestAccessor;
+using RequestDispatcherOldAccessor = DB::KeeperRequestDispatcherOldTestAccessor;
+using DispatcherAccessor = DB::KeeperDispatcherTestAccessor;
+
+}
+
+/// A SessionID commit from another server must not retire our in-flight SessionID request, and our
+/// own must still retire it.
+TEST(KeeperDispatcher, SessionIDCommitCorrelation)
+{
+    DispatcherFixture fixture;
+    auto & dispatcher = *fixture.dispatcher;
+
+    size_t batch_idx = RequestDispatcherAccessor::seedInFlightBatch(
+        dispatcher, makeSessionIDRequest(/*server_id=*/ 1, /*internal_id=*/ 7));
+
+    ASSERT_EQ(RequestDispatcherAccessor::committedRequests(dispatcher, batch_idx), 0u);
+
+    /// Same degenerate (session_id, xid), different origin.
+    dispatcher.onCommit(makeSessionIDRequest(/*server_id=*/ 2, /*internal_id=*/ 7));
+    EXPECT_EQ(RequestDispatcherAccessor::committedRequests(dispatcher, batch_idx), 0u)
+        << "a foreign server's SessionID commit retired our request";
+
+    /// Same server, different client.
+    dispatcher.onCommit(makeSessionIDRequest(/*server_id=*/ 1, /*internal_id=*/ 8));
+    EXPECT_EQ(RequestDispatcherAccessor::committedRequests(dispatcher, batch_idx), 0u)
+        << "another client's SessionID commit retired our request";
+
+    /// Ours: correlation must still work, otherwise every session request would stall.
+    dispatcher.onCommit(makeSessionIDRequest(/*server_id=*/ 1, /*internal_id=*/ 7));
+    EXPECT_EQ(RequestDispatcherAccessor::committedRequests(dispatcher, batch_idx), 1u)
+        << "our own SessionID commit did not retire our request";
+    EXPECT_EQ(RequestDispatcherAccessor::headIdx(dispatcher), batch_idx + 1) << "the fully committed batch was not popped";
+}
+
+/// A dropped SessionID request must reach its waiter instead of the per-session response queue,
+/// where session id -1 has no callback and the response is discarded.
+TEST(KeeperDispatcher, SessionIDErrorReachesWaiter)
+{
+    DispatcherFixture fixture;
+    auto & dispatcher = *fixture.dispatcher;
+
+    size_t batch_idx = RequestDispatcherAccessor::seedInFlightBatch(
+        dispatcher, makeSessionIDRequest(/*server_id=*/ 1, /*internal_id=*/ 11));
+
+    RequestDispatcherAccessor::dropInFlightRequests(dispatcher);
+
+    ASSERT_EQ(fixture.routed.size(), 1u) << "the dropped SessionID error did not reach its waiter";
+    const auto & response = fixture.routed.front();
+    EXPECT_EQ(response.response->error, Coordination::Error::ZCONNECTIONLOSS);
+
+    /// The identifiers the waiter is keyed by must survive makeResponse().
+    const auto & session_id_response = dynamic_cast<const Coordination::ZooKeeperSessionIDResponse &>(*response.response);
+    EXPECT_EQ(session_id_response.server_id, 1);
+    EXPECT_EQ(session_id_response.internal_id, 11);
+
+    EXPECT_EQ(RequestDispatcherAccessor::headIdx(dispatcher), batch_idx + 1) << "the dropped batch was not popped";
+}
+
+/// use_new_dispatcher is a setting, so the old dispatcher is a live carrier of the same defect. It
+/// has no in-flight batch tracking, so it synthesizes the error straight from addErrorResponses.
+TEST(KeeperDispatcherOld, SessionIDErrorReachesWaiter)
+{
+    DispatcherFixture fixture;
+    DB::KeeperRequestDispatcherOld dispatcher_old(fixture.server.get(), fixture.router());
+    /// Its threads loop until the context says shutdown, the way KeeperDispatcher::shutdown does it.
+    SCOPE_EXIT({
+        fixture.keeper_context->setShutdownCalled();
+        dispatcher_old.shutdown();
+    });
+
+    RequestDispatcherOldAccessor::addErrorResponses(
+        dispatcher_old,
+        {makeSessionIDRequest(/*server_id=*/ 1, /*internal_id=*/ 13)},
+        Coordination::Error::ZCONNECTIONLOSS,
+        /*may_have_dependent_reads=*/ false);
+
+    ASSERT_EQ(fixture.routed.size(), 1u) << "the dropped SessionID error did not reach its waiter";
+    const auto & response = fixture.routed.front();
+    EXPECT_EQ(response.response->error, Coordination::Error::ZCONNECTIONLOSS);
+    const auto & session_id_response = dynamic_cast<const Coordination::ZooKeeperSessionIDResponse &>(*response.response);
+    EXPECT_EQ(session_id_response.server_id, 1);
+    EXPECT_EQ(session_id_response.internal_id, 13);
+}
+
+/// Unlike the arms above, which stop at the router seam, this one drives the production router and
+/// asserts on the getSessionID waiter a client actually blocks on.
+TEST(KeeperDispatcher, SessionIDErrorReachesRealWaiter)
+{
+    DispatcherFixture fixture;
+
+    /// onSessionIDResponse reads only server->getServerID(), set by the KeeperServer constructor, so
+    /// an un-started server suffices here.
+    DB::KeeperDispatcher keeper_dispatcher;
+    /// Holds a raw pointer to the server keeper_dispatcher takes over, and would outlive it.
+    fixture.dispatcher.reset();
+    DispatcherAccessor::setServer(keeper_dispatcher, std::move(fixture.server));
+
+    DB::KeeperRequestDispatcher dispatcher(
+        DispatcherAccessor::server(keeper_dispatcher), DispatcherAccessor::router(keeper_dispatcher));
+
+    constexpr int64_t internal_id = 17;
+    auto waiter = DispatcherAccessor::registerSessionIDWaiter(keeper_dispatcher, internal_id);
+    ASSERT_TRUE(waiter.has_value());
+    auto & future = *waiter;
+
+    /// A response for a different client must not wake our waiter.
+    RequestDispatcherAccessor::seedInFlightBatch(dispatcher, makeSessionIDRequest(/*server_id=*/ 1, /*internal_id=*/ 18));
+    RequestDispatcherAccessor::dropInFlightRequests(dispatcher);
+    ASSERT_EQ(future.wait_for(std::chrono::seconds(0)), std::future_status::timeout)
+        << "another client's SessionID error woke our waiter";
+    EXPECT_EQ(DispatcherAccessor::sessionIDWaiterCount(keeper_dispatcher, internal_id), 1u);
+
+    RequestDispatcherAccessor::seedInFlightBatch(dispatcher, makeSessionIDRequest(/*server_id=*/ 1, internal_id));
+    RequestDispatcherAccessor::dropInFlightRequests(dispatcher);
+
+    /// Ready without waiting: the client does not sit out the session timeout.
+    ASSERT_EQ(future.wait_for(std::chrono::seconds(0)), std::future_status::ready)
+        << "the dropped SessionID error did not reach the getSessionID waiter";
+
+    try
+    {
+        FAIL() << "getSessionID returned session id " << future.get() << " instead of the error";
+    }
+    catch (const Coordination::Exception & e)
+    {
+        EXPECT_EQ(e.code, Coordination::Error::ZCONNECTIONLOSS);
+    }
+
+    EXPECT_EQ(DispatcherAccessor::sessionIDWaiterCount(keeper_dispatcher, internal_id), 0u) << "the waiter entry leaked";
+}
+
+/// A request accepted before the shutdown flag was set is discarded without a response: the drains
+/// do not synthesize one, and the old dispatcher's requestThread abandons what it already popped.
+/// Routing cannot cover that, so the waiter is completed once no dispatcher can produce a response.
+TEST(KeeperDispatcher, PendingSessionIDRequestsFailOnShutdown)
+{
+    DispatcherFixture fixture;
+    fixture.dispatcher.reset();
+
+    /// Drives the real KeeperDispatcher::shutdown, so removing its call reddens this arm. Neither
+    /// dispatcher is constructed and `server` is left null: shutdown skips both, and the waiter
+    /// must still be completed.
+    DB::KeeperDispatcher keeper_dispatcher;
+    DispatcherAccessor::setKeeperContext(keeper_dispatcher, fixture.keeper_context);
+
+    constexpr int64_t internal_id = 31;
+    auto waiter = DispatcherAccessor::registerSessionIDWaiter(keeper_dispatcher, internal_id);
+    ASSERT_TRUE(waiter.has_value());
+    auto & future = *waiter;
+
+    ASSERT_EQ(future.wait_for(std::chrono::seconds(0)), std::future_status::timeout)
+        << "the waiter completed before shutdown";
+
+    keeper_dispatcher.shutdown(/*closed_all_connections=*/ true);
+
+    ASSERT_EQ(future.wait_for(std::chrono::seconds(0)), std::future_status::ready)
+        << "the waiter was left to time out across shutdown";
+    try
+    {
+        FAIL() << "getSessionID returned session id " << future.get() << " instead of the error";
+    }
+    catch (const Coordination::Exception & e)
+    {
+        EXPECT_EQ(e.code, Coordination::Error::ZSESSIONEXPIRED);
+    }
+
+    EXPECT_EQ(DispatcherAccessor::sessionIDWaiterCount(keeper_dispatcher, internal_id), 0u) << "the waiter entry leaked";
+}
+
+/// setShutdownCalled is one-shot, so a shutdown step that throws before the normal cleanup point
+/// is the waiters' only chance to be completed.
+TEST(KeeperDispatcher, PendingSessionIDRequestsFailOnThrowingShutdown)
+{
+    DispatcherFixture fixture;
+    fixture.dispatcher.reset();
+
+    DB::KeeperDispatcher keeper_dispatcher;
+    DispatcherAccessor::setKeeperContext(keeper_dispatcher, fixture.keeper_context);
+
+    constexpr int64_t internal_id = 37;
+    auto waiter = DispatcherAccessor::registerSessionIDWaiter(keeper_dispatcher, internal_id);
+    ASSERT_TRUE(waiter.has_value());
+    auto & future = *waiter;
+
+    DB::FailPointInjection::enableFailPoint("keeper_shutdown_throw_after_flag");
+    SCOPE_EXIT({ DB::FailPointInjection::disableFailPoint("keeper_shutdown_throw_after_flag"); });
+
+    keeper_dispatcher.shutdown(/*closed_all_connections=*/ true);
+
+    ASSERT_EQ(future.wait_for(std::chrono::seconds(0)), std::future_status::ready)
+        << "a shutdown that threw left the waiter to time out";
+    try
+    {
+        FAIL() << "getSessionID returned session id " << future.get() << " instead of the error";
+    }
+    catch (const Coordination::Exception & e)
+    {
+        EXPECT_EQ(e.code, Coordination::Error::ZSESSIONEXPIRED);
+    }
+
+    EXPECT_EQ(DispatcherAccessor::sessionIDWaiterCount(keeper_dispatcher, internal_id), 0u) << "the waiter entry leaked";
 }
 
 #endif
